refactor: encapsulate levels data fetching logic into useLevelsFetch hook ss-057
