+++ src/libs/hooks/hooks.ts
@@ -4,6 +4,7 @@ export { useAudioPlayer } from "./use-audio-player/use-audio-player.hook";
 export { useStopAudioOnUnmount } from "./use-audio-player/use-stop-audio-on-unmount.hook";
 export { useGameFetch } from "./use-game-fetch/use-game-fetch.hook";
 export { useLanguageSync } from "./use-language-sync/use-language-sync.hook";
+export { useLevelsFetch } from "./use-levels-fetch/use-levels-fetch.hook";
 export { useTrueFalseGame } from "./use-true-false-game/use-true-false-game.hook";
 export { useCallback, useEffect, useId, useMemo, useRef, useState } from "react";
 export { useForm } from "react-hook-form";
