+++ src/pages/game-content-page/game-content-page.tsx
@@ -1,17 +1,11 @@
-import { useCallback, useEffect } from "react";
-
 import { FallbackMessage, Loader } from "~/libs/components/components";
 import { GameLevelsPreview } from "~/libs/components/game-levels-preview/game-levels-preview";
-import { DataStatus } from "~/libs/enums/enums";
 import {
-	useAppDispatch,
-	useAppSelector,
 	useGameFetch,
-	useLanguageSync,
+	useLevelsFetch,
 	useParams,
 	useTranslation,
 } from "~/libs/hooks/hooks";
-import { getLevelsList } from "~/modules/games/slices/actions";
 
 import styles from "./styles.module.css";
 
@@ -19,23 +13,7 @@ const GameContentPage: React.FC = () => {
 	const { t } = useTranslation();
 	const { id } = useParams();
 	const { currentGame, isLoading: isGameLoading } = useGameFetch(id);
-
-	const dispatch = useAppDispatch();
-	const levelsStatus = useAppSelector((state) => state.games.levelsStatus);
-
-	const fetchLevels = useCallback(() => {
-		if (id) {
-			void dispatch(getLevelsList(id));
-		}
-	}, [dispatch, id]);
-
-	useLanguageSync(fetchLevels);
-
-	useEffect(() => {
-		if (id && levelsStatus === DataStatus.IDLE) {
-			fetchLevels();
-		}
-	}, [id, fetchLevels, levelsStatus]);
+	const { isLoading: isLevelsLoading } = useLevelsFetch(id);
 
 	const renderError = (message: string): React.JSX.Element => (
 		<div className={styles["game-content-page__error-container"]}>
@@ -52,8 +30,6 @@ const GameContentPage: React.FC = () => {
 		return renderError(t("games.content.invalidId"));
 	}
 
-	const isLevelsLoading =
-		levelsStatus === DataStatus.PENDING || levelsStatus === DataStatus.IDLE;
 	const isPageLoading = isGameLoading || isLevelsLoading;
 
 	if (isPageLoading) {
